@@ -1,6 +1,6 @@
 [metadata]
 name = cqlsh-expansion
-version = 0.7.8
+version = 0.8.9
 description = The cqlsh-expansion utility extends native cqlsh functionality to include parameters and capabilities specific to Amazon Keyspaces such as support for Sigv4 Authentication.
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -16,9 +16,12 @@ classifiers =
 	Intended Audience :: Developers
 	Intended Audience :: End Users/Desktop
 	Programming Language :: Python
-	Programming Language :: Python :: 2
-	Programming Language :: Python :: 2.7
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3.6
+    Programming Language :: Python :: 3.7
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
 
 [options]
-python_requires = ~= 2.7
+python_requires = >= 3.6
 scripts = bin/cqlsh-expansion, bin/cqlsh-expansion.py, bin/cqlsh, bin/cqlsh.py
