@@ -14,15 +14,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .cqlhandling import CqlParsingRuleSet, Hint
 from cassandra.metadata import maybe_escape_name
+from cqlshlib import helptopics
+from cqlshlib.cqlhandling import CqlParsingRuleSet, Hint
 
-
-simple_cql_types = set(('ascii', 'bigint', 'blob', 'boolean', 'counter', 'date', 'decimal', 'double', 'duration', 'float',
-                        'inet', 'int', 'smallint', 'text', 'time', 'timestamp', 'timeuuid', 'tinyint', 'uuid', 'varchar', 'varint'))
+simple_cql_types = {'ascii', 'bigint', 'blob', 'boolean', 'counter', 'date', 'decimal', 'double', 'duration', 'float',
+                    'inet', 'int', 'smallint', 'text', 'time', 'timestamp', 'timeuuid', 'tinyint', 'uuid', 'varchar',
+                    'varint'}
 simple_cql_types.difference_update(('set', 'map', 'list'))
 
-from . import helptopics
 cqldocs = helptopics.CQL3HelpTopics()
 
 
@@ -35,8 +35,9 @@ class UnexpectedTableStructure(UserWarning):
         return 'Unexpected table structure; may not translate correctly to CQL. ' + self.msg
 
 
-SYSTEM_KEYSPACES = ('system', 'system_schema', 'system_traces', 'system_auth', 'system_distributed')
-NONALTERBALE_KEYSPACES = ('system', 'system_schema')
+SYSTEM_KEYSPACES = ('system', 'system_schema', 'system_traces', 'system_auth', 'system_distributed', 'system_views',
+                    'system_virtual_schema')
+NONALTERBALE_KEYSPACES = ('system', 'system_schema', 'system_views', 'system_virtual_schema')
 
 
 class Cql3ParsingRuleSet(CqlParsingRuleSet):
@@ -44,22 +45,23 @@ class Cql3ParsingRuleSet(CqlParsingRuleSet):
     columnfamily_layout_options = (
         ('bloom_filter_fp_chance', None),
         ('comment', None),
-        ('dclocal_read_repair_chance', 'local_read_repair_chance'),
         ('gc_grace_seconds', None),
         ('min_index_interval', None),
         ('max_index_interval', None),
-        ('read_repair_chance', None),
         ('default_time_to_live', None),
         ('speculative_retry', None),
+        ('additional_write_policy', None),
         ('memtable_flush_period_in_ms', None),
-        ('cdc', None)
+        ('cdc', None),
+        ('read_repair', None),
     )
 
     columnfamily_layout_map_options = (
         # (CQL3 option name, schema_columnfamilies column name (or None if same),
         #  list of known map keys)
         ('compaction', 'compaction_strategy_options',
-            ('class', 'max_threshold', 'tombstone_compaction_interval', 'tombstone_threshold', 'enabled', 'unchecked_tombstone_compaction', 'only_purge_repaired_tombstones')),
+            ('class', 'max_threshold', 'tombstone_compaction_interval', 'tombstone_threshold', 'enabled',
+             'unchecked_tombstone_compaction', 'only_purge_repaired_tombstones', 'provide_overlapping_tombstones')),
         ('compression', 'compression_parameters',
             ('sstable_compression', 'chunk_length_kb', 'crc_check_chance')),
         ('caching', None,
@@ -78,6 +80,33 @@ class Cql3ParsingRuleSet(CqlParsingRuleSet):
         'LOCAL_QUORUM',
         'EACH_QUORUM',
         'SERIAL'
+    )
+
+    size_tiered_compaction_strategy_options = (
+        'min_sstable_size',
+        'min_threshold',
+        'bucket_high',
+        'bucket_low'
+    )
+
+    leveled_compaction_strategy_options = (
+        'sstable_size_in_mb',
+        'fanout_size'
+    )
+
+    date_tiered_compaction_strategy_options = (
+        'base_time_seconds',
+        'max_sstable_age_days',
+        'min_threshold',
+        'max_window_size_seconds',
+        'timestamp_resolution'
+    )
+
+    time_window_compaction_strategy_options = (
+        'compaction_window_unit',
+        'compaction_window_size',
+        'min_threshold',
+        'timestamp_resolution'
     )
 
     @classmethod
@@ -154,7 +183,7 @@ JUNK ::= /([ \t\r\f\v]+|(--|[/][/])[^\n\r]*([\n\r]|$)|[/][*].*?[*][/])/ ;
 <colon> ::=         ":" ;
 <star> ::=          "*" ;
 <endtoken> ::=      ";" ;
-<op> ::=            /[-+=,().]/ ;
+<op> ::=            /[-+=%/,().]/ ;
 <cmp> ::=           /[<>!]=?/ ;
 <brackets> ::=      /[][{}]/ ;
 
@@ -359,6 +388,11 @@ completer_for('property', 'propeq')(prop_equals_completer)
 def prop_name_completer(ctxt, cass):
     if working_on_keyspace(ctxt):
         return ks_prop_name_completer(ctxt, cass)
+    elif 'MATERIALIZED' == ctxt.get_binding('wat', '').upper():
+        props = cf_prop_name_completer(ctxt, cass)
+        props.remove('default_time_to_live')
+        props.remove('gc_grace_seconds')
+        return props
     else:
         return cf_prop_name_completer(ctxt, cass)
 
@@ -415,19 +449,19 @@ def ks_prop_val_mapkey_completer(ctxt, cass):
     optname = ctxt.get_binding('propname')[-1]
     if optname != 'replication':
         return ()
-    keysseen = map(dequote_value, ctxt.get_binding('propmapkey', ()))
-    valsseen = map(dequote_value, ctxt.get_binding('propmapval', ()))
+    keysseen = list(map(dequote_value, ctxt.get_binding('propmapkey', ())))
+    valsseen = list(map(dequote_value, ctxt.get_binding('propmapval', ())))
     for k, v in zip(keysseen, valsseen):
         if k == 'class':
             repclass = v
             break
     else:
         return ["'class'"]
-    if repclass in CqlRuleSet.replication_factor_strategies:
-        opts = set(('replication_factor',))
+    if repclass == 'SimpleStrategy':
+        opts = {'replication_factor'}
     elif repclass == 'NetworkTopologyStrategy':
         return [Hint('<dc_name>')]
-    return map(escape_value, opts.difference(keysseen))
+    return list(map(escape_value, opts.difference(keysseen)))
 
 
 def ks_prop_val_mapval_completer(ctxt, cass):
@@ -436,7 +470,7 @@ def ks_prop_val_mapval_completer(ctxt, cass):
         return ()
     currentkey = dequote_value(ctxt.get_binding('propmapkey')[-1])
     if currentkey == 'class':
-        return map(escape_value, CqlRuleSet.replication_strategies)
+        return list(map(escape_value, CqlRuleSet.replication_strategies))
     return [Hint('<term>')]
 
 
@@ -444,15 +478,15 @@ def ks_prop_val_mapender_completer(ctxt, cass):
     optname = ctxt.get_binding('propname')[-1]
     if optname != 'replication':
         return [',']
-    keysseen = map(dequote_value, ctxt.get_binding('propmapkey', ()))
-    valsseen = map(dequote_value, ctxt.get_binding('propmapval', ()))
+    keysseen = list(map(dequote_value, ctxt.get_binding('propmapkey', ())))
+    valsseen = list(map(dequote_value, ctxt.get_binding('propmapval', ())))
     for k, v in zip(keysseen, valsseen):
         if k == 'class':
             repclass = v
             break
     else:
         return [',']
-    if repclass in CqlRuleSet.replication_factor_strategies:
+    if repclass == 'SimpleStrategy':
         if 'replication_factor' not in keysseen:
             return [',']
     if repclass == 'NetworkTopologyStrategy' and len(keysseen) == 1:
@@ -461,8 +495,8 @@ def ks_prop_val_mapender_completer(ctxt, cass):
 
 
 def cf_prop_name_completer(ctxt, cass):
-    return [c[0] for c in (CqlRuleSet.columnfamily_layout_options +
-                           CqlRuleSet.columnfamily_layout_map_options)]
+    return [c[0] for c in (CqlRuleSet.columnfamily_layout_options
+                           + CqlRuleSet.columnfamily_layout_map_options)]
 
 
 def cf_prop_val_completer(ctxt, cass):
@@ -476,14 +510,15 @@ def cf_prop_val_completer(ctxt, cass):
         return ["{'keys': '"]
     if any(this_opt == opt[0] for opt in CqlRuleSet.obsolete_cf_options):
         return ["'<obsolete_option>'"]
-    if this_opt in ('read_repair_chance', 'bloom_filter_fp_chance',
-                    'dclocal_read_repair_chance'):
+    if this_opt == 'bloom_filter_fp_chance':
         return [Hint('<float_between_0_and_1>')]
     if this_opt in ('min_compaction_threshold', 'max_compaction_threshold',
                     'gc_grace_seconds', 'min_index_interval', 'max_index_interval'):
         return [Hint('<integer>')]
     if this_opt in ('cdc'):
         return [Hint('<true|false>')]
+    if this_opt in ('read_repair'):
+        return [Hint('<\'none\'|\'blocking\'>')]
     return [Hint('<option_value>')]
 
 
@@ -494,13 +529,13 @@ def cf_prop_val_mapkey_completer(ctxt, cass):
             break
     else:
         return ()
-    keysseen = map(dequote_value, ctxt.get_binding('propmapkey', ()))
-    valsseen = map(dequote_value, ctxt.get_binding('propmapval', ()))
-    pairsseen = dict(zip(keysseen, valsseen))
+    keysseen = list(map(dequote_value, ctxt.get_binding('propmapkey', ())))
+    valsseen = list(map(dequote_value, ctxt.get_binding('propmapval', ())))
+    pairsseen = dict(list(zip(keysseen, valsseen)))
     if optname == 'compression':
-        return map(escape_value, set(subopts).difference(keysseen))
+        return list(map(escape_value, set(subopts).difference(keysseen)))
     if optname == 'caching':
-        return map(escape_value, set(subopts).difference(keysseen))
+        return list(map(escape_value, set(subopts).difference(keysseen)))
     if optname == 'compaction':
         opts = set(subopts)
         try:
@@ -509,27 +544,15 @@ def cf_prop_val_mapkey_completer(ctxt, cass):
             return ["'class'"]
         csc = csc.split('.')[-1]
         if csc == 'SizeTieredCompactionStrategy':
-            opts.add('min_sstable_size')
-            opts.add('min_threshold')
-            opts.add('bucket_high')
-            opts.add('bucket_low')
+            opts = opts.union(set(CqlRuleSet.size_tiered_compaction_strategy_options))
         elif csc == 'LeveledCompactionStrategy':
-            opts.add('sstable_size_in_mb')
-            opts.add('fanout_size')
+            opts = opts.union(set(CqlRuleSet.leveled_compaction_strategy_options))
         elif csc == 'DateTieredCompactionStrategy':
-            opts.add('base_time_seconds')
-            opts.add('max_sstable_age_days')
-            opts.add('min_threshold')
-            opts.add('max_window_size_seconds')
-            opts.add('timestamp_resolution')
+            opts = opts.union(set(CqlRuleSet.date_tiered_compaction_strategy_options))
         elif csc == 'TimeWindowCompactionStrategy':
-            opts.add('compaction_window_unit')
-            opts.add('compaction_window_size')
-            opts.add('min_threshold')
-            opts.add('max_threshold')
-            opts.add('timestamp_resolution')
+            opts = opts.union(set(CqlRuleSet.time_window_compaction_strategy_options))
 
-        return map(escape_value, opts)
+        return list(map(escape_value, opts))
     return ()
 
 
@@ -538,11 +561,13 @@ def cf_prop_val_mapval_completer(ctxt, cass):
     key = dequote_value(ctxt.get_binding('propmapkey')[-1])
     if opt == 'compaction':
         if key == 'class':
-            return map(escape_value, CqlRuleSet.available_compaction_classes)
+            return list(map(escape_value, CqlRuleSet.available_compaction_classes))
+        if key == 'provide_overlapping_tombstones':
+            return [Hint('<NONE|ROW|CELL>')]
         return [Hint('<option_value>')]
     elif opt == 'compression':
         if key == 'sstable_compression':
-            return map(escape_value, CqlRuleSet.available_compression_classes)
+            return list(map(escape_value, CqlRuleSet.available_compression_classes))
         return [Hint('<option_value>')]
     elif opt == 'caching':
         if key == 'rows_per_partition':
@@ -568,19 +593,19 @@ def storagetype_completer(ctxt, cass):
 
 @completer_for('keyspaceName', 'ksname')
 def ks_name_completer(ctxt, cass):
-    return map(maybe_escape_name, cass.get_keyspace_names())
+    return list(map(maybe_escape_name, cass.get_keyspace_names()))
 
 
 @completer_for('nonSystemKeyspaceName', 'ksname')
 def non_system_ks_name_completer(ctxt, cass):
     ksnames = [n for n in cass.get_keyspace_names() if n not in SYSTEM_KEYSPACES]
-    return map(maybe_escape_name, ksnames)
+    return list(map(maybe_escape_name, ksnames))
 
 
 @completer_for('alterableKeyspaceName', 'ksname')
 def alterable_ks_name_completer(ctxt, cass):
     ksnames = [n for n in cass.get_keyspace_names() if n not in NONALTERBALE_KEYSPACES]
-    return map(maybe_escape_name, ksnames)
+    return list(map(maybe_escape_name, ksnames))
 
 
 def cf_ks_name_completer(ctxt, cass):
@@ -613,7 +638,7 @@ def cf_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, cfnames)
+    return list(map(maybe_escape_name, cfnames))
 
 
 @completer_for('materializedViewName', 'mvname')
@@ -627,7 +652,7 @@ def mv_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, mvnames)
+    return list(map(maybe_escape_name, mvnames))
 
 
 completer_for('userTypeName', 'ksname')(cf_ks_name_completer)
@@ -645,7 +670,7 @@ def ut_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, utnames)
+    return list(map(maybe_escape_name, utnames))
 
 
 completer_for('userTypeName', 'utname')(ut_name_completer)
@@ -677,7 +702,7 @@ def get_ut_layout(ctxt, cass):
 
 
 def working_on_keyspace(ctxt):
-    wat = ctxt.get_binding('wat').upper()
+    wat = ctxt.get_binding('wat', '').upper()
     if wat in ('KEYSPACE', 'SCHEMA'):
         return True
     return False
@@ -708,7 +733,7 @@ syntax_rules += r'''
                  ;
 <udtSubfieldSelection> ::= <identifier> "." <identifier>
                          ;
-<selector> ::= [colname]=<cident>
+<selector> ::= [colname]=<cident> ( "[" ( <term> ( ".." <term> "]" )? | <term> ".." ) )?
              | <udtSubfieldSelection>
              | "WRITETIME" "(" [colname]=<cident> ")"
              | "TTL" "(" [colname]=<cident> ")"
@@ -722,7 +747,13 @@ syntax_rules += r'''
 <orderByClause> ::= [ordercol]=<cident> ( "ASC" | "DESC" )?
                   ;
 <groupByClause> ::= [groupcol]=<cident>
+                  | <functionName><groupByFunctionArguments>
                   ;
+<groupByFunctionArguments> ::= "(" ( <groupByFunctionArgument> ( "," <groupByFunctionArgument> )* )? ")"
+                             ;
+<groupByFunctionArgument> ::= [groupcol]=<cident>
+                            | <term>
+                            ;
 '''
 
 
@@ -736,7 +767,7 @@ def udf_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, udfnames)
+    return list(map(maybe_escape_name, udfnames))
 
 
 def uda_name_completer(ctxt, cass):
@@ -749,7 +780,7 @@ def uda_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, udanames)
+    return list(map(maybe_escape_name, udanames))
 
 
 def udf_uda_name_completer(ctxt, cass):
@@ -762,7 +793,7 @@ def udf_uda_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, functionnames)
+    return list(map(maybe_escape_name, functionnames))
 
 
 def ref_udf_name_completer(ctxt, cass):
@@ -770,7 +801,7 @@ def ref_udf_name_completer(ctxt, cass):
         udanames = cass.get_userfunction_names(None)
     except Exception:
         return ()
-    return map(maybe_escape_name, udanames)
+    return list(map(maybe_escape_name, udanames))
 
 
 completer_for('functionAggregateName', 'ksname')(cf_ks_name_completer)
@@ -828,7 +859,7 @@ def relation_token_subject_completer(ctxt, cass):
 def select_relation_lhs_completer(ctxt, cass):
     layout = get_table_meta(ctxt, cass)
     filterable = set()
-    already_filtered_on = map(dequote_name, ctxt.get_binding('rel_lhs', ()))
+    already_filtered_on = list(map(dequote_name, ctxt.get_binding('rel_lhs', ())))
     for num in range(0, len(layout.partition_key)):
         if num == 0 or layout.partition_key[num - 1].name in already_filtered_on:
             filterable.add(layout.partition_key[num].name)
@@ -839,9 +870,9 @@ def select_relation_lhs_completer(ctxt, cass):
             filterable.add(layout.clustering_key[num].name)
         else:
             break
-    for idx in layout.indexes.itervalues():
+    for idx in layout.indexes.values():
         filterable.add(idx.index_options["target"])
-    return map(maybe_escape_name, filterable)
+    return list(map(maybe_escape_name, filterable))
 
 
 explain_completion('selector', 'colname')
@@ -864,9 +895,9 @@ syntax_rules += r'''
 def regular_column_names(table_meta):
     if not table_meta or not table_meta.columns:
         return []
-    regular_columns = list(set(table_meta.columns.keys()) -
-                           set([key.name for key in table_meta.partition_key]) -
-                           set([key.name for key in table_meta.clustering_key]))
+    regular_columns = list(set(table_meta.columns.keys())
+                           - set([key.name for key in table_meta.partition_key])
+                           - set([key.name for key in table_meta.clustering_key]))
     return regular_columns
 
 
@@ -879,13 +910,13 @@ def insert_colname_completer(ctxt, cass):
         if k.name not in colnames:
             return [maybe_escape_name(k.name)]
     normalcols = set(regular_column_names(layout)) - colnames
-    return map(maybe_escape_name, normalcols)
+    return list(map(maybe_escape_name, normalcols))
 
 
 @completer_for('insertStatement', 'newval')
 def insert_newval_completer(ctxt, cass):
     layout = get_table_meta(ctxt, cass)
-    insertcols = map(dequote_name, ctxt.get_binding('colname'))
+    insertcols = list(map(dequote_name, ctxt.get_binding('colname')))
     valuesdone = ctxt.get_binding('newval', ())
     if len(valuesdone) >= len(insertcols):
         return []
@@ -936,7 +967,7 @@ syntax_rules += r'''
                ;
 <conditions> ::=  <condition> ( "AND" <condition> )*
                ;
-<condition_op_and_rhs> ::= (("=" | "<" | ">" | "<=" | ">=" | "!=") <term>)
+<condition_op_and_rhs> ::= (("=" | "<" | ">" | "<=" | ">=" | "!=" | "CONTAINS" ( "KEY" )? ) <term>)
                            | ("IN" "(" <term> ( "," <term> )* ")" )
                          ;
 <condition> ::= conditioncol=<cident>
@@ -958,7 +989,7 @@ def update_option_completer(ctxt, cass):
 @completer_for('assignment', 'updatecol')
 def update_col_completer(ctxt, cass):
     layout = get_table_meta(ctxt, cass)
-    return map(maybe_escape_name, regular_column_names(layout))
+    return list(map(maybe_escape_name, regular_column_names(layout)))
 
 
 @completer_for('assignment', 'update_rhs')
@@ -1100,7 +1131,7 @@ def delete_opt_completer(ctxt, cass):
 @completer_for('deleteSelector', 'delcol')
 def delete_delcol_completer(ctxt, cass):
     layout = get_table_meta(ctxt, cass)
-    return map(maybe_escape_name, regular_column_names(layout))
+    return list(map(maybe_escape_name, regular_column_names(layout)))
 
 
 syntax_rules += r'''
@@ -1181,7 +1212,7 @@ syntax_rules += r'''
 
 @completer_for('cfamOrdering', 'ordercol')
 def create_cf_clustering_order_colname_completer(ctxt, cass):
-    colnames = map(dequote_name, ctxt.get_binding('newcolname', ()))
+    colnames = list(map(dequote_name, ctxt.get_binding('newcolname', ())))
     # Definitely some of these aren't valid for ordering, but I'm not sure
     # precisely which are. This is good enough for now
     return colnames
@@ -1211,7 +1242,7 @@ def create_cf_ks_dot_completer(ctxt, cass):
 def create_cf_pkdef_declaration_completer(ctxt, cass):
     cols_declared = ctxt.get_binding('newcolname')
     pieces_already = ctxt.get_binding('ptkey', ())
-    pieces_already = map(dequote_name, pieces_already)
+    pieces_already = list(map(dequote_name, pieces_already))
     while cols_declared[0] in pieces_already:
         cols_declared = cols_declared[1:]
         if len(cols_declared) < 2:
@@ -1223,7 +1254,7 @@ def create_cf_pkdef_declaration_completer(ctxt, cass):
 def create_cf_composite_key_declaration_completer(ctxt, cass):
     cols_declared = ctxt.get_binding('newcolname')
     pieces_already = ctxt.get_binding('ptkey', ()) + ctxt.get_binding('pkey', ())
-    pieces_already = map(dequote_name, pieces_already)
+    pieces_already = list(map(dequote_name, pieces_already))
     while cols_declared[0] in pieces_already:
         cols_declared = cols_declared[1:]
         if len(cols_declared) < 2:
@@ -1265,9 +1296,16 @@ syntax_rules += r'''
                                ( "USING" <stringLiteral> ( "WITH" "OPTIONS" "=" <mapLiteral> )? )?
                          ;
 
-<createMaterializedViewStatement> ::= "CREATE" "MATERIALIZED" "VIEW" ("IF" "NOT" "EXISTS")? <materializedViewName>?
-                                      "AS" <selectStatement>
-                                      "PRIMARY" "KEY" <pkDef>
+
+<colList> ::= "(" <cident> ( "," <cident> )* ")"
+          ;
+
+<createMaterializedViewStatement> ::= "CREATE" wat="MATERIALIZED" "VIEW" ("IF" "NOT" "EXISTS")? viewname=<materializedViewName>?
+                                      "AS" "SELECT" <selectClause>
+                                      "FROM" cf=<columnFamilyName>
+                                      "WHERE" <cident> "IS" "NOT" "NULL" ( "AND" <cident> "IS" "NOT" "NULL")*
+                                      "PRIMARY" "KEY" (<colList> | ( "(" <colList> ( "," <cident> )* ")" ))
+                                      ( "WITH" <cfamProperty> ( "AND" <cfamProperty> )* )?
                                     ;
 
 <createUserTypeStatement> ::= "CREATE" "TYPE" ( ks=<nonSystemKeyspaceName> dot="." )? typename=<cfOrKsName> "(" newcol=<cident> <storageType>
@@ -1301,6 +1339,7 @@ syntax_rules += r'''
 '''
 
 explain_completion('createIndexStatement', 'indexname', '<new_index_name>')
+explain_completion('createMaterializedViewStatement', 'viewname', '<new_view_name>')
 explain_completion('createUserTypeStatement', 'typename', '<new_type_name>')
 explain_completion('createUserTypeStatement', 'newcol', '<new_field_name>')
 
@@ -1309,9 +1348,9 @@ explain_completion('createUserTypeStatement', 'newcol', '<new_field_name>')
 def create_index_col_completer(ctxt, cass):
     """ Return the columns for which an index doesn't exist yet. """
     layout = get_table_meta(ctxt, cass)
-    idx_targets = [idx.index_options["target"] for idx in layout.indexes.itervalues()]
-    colnames = [cd.name for cd in layout.columns.values() if cd.name not in idx_targets]
-    return map(maybe_escape_name, colnames)
+    idx_targets = [idx.index_options["target"] for idx in layout.indexes.values()]
+    colnames = [cd.name for cd in list(layout.columns.values()) if cd.name not in idx_targets]
+    return list(map(maybe_escape_name, colnames))
 
 
 syntax_rules += r'''
@@ -1369,25 +1408,25 @@ def idx_ks_idx_name_completer(ctxt, cass):
         if ks is None:
             return ()
         raise
-    return map(maybe_escape_name, idxnames)
+    return list(map(maybe_escape_name, idxnames))
 
 
 syntax_rules += r'''
-<alterTableStatement> ::= "ALTER" wat=( "COLUMNFAMILY" | "TABLE" ) cf=<columnFamilyName>
+<alterTableStatement> ::= "ALTER" wat=( "COLUMNFAMILY" | "TABLE" ) ("IF" "EXISTS")? cf=<columnFamilyName>
                                <alterInstructions>
                         ;
-<alterInstructions> ::= "ADD" newcol=<cident> <storageType> ("static")?
-                      | "DROP" existcol=<cident>
+<alterInstructions> ::= "ADD" ("IF" "NOT" "EXISTS")? newcol=<cident> <storageType> ("static")?
+                      | "DROP" ("IF" "EXISTS")? existcol=<cident>
                       | "WITH" <cfamProperty> ( "AND" <cfamProperty> )*
-                      | "RENAME" existcol=<cident> "TO" newcol=<cident>
+                      | "RENAME" ("IF" "EXISTS")? existcol=<cident> "TO" newcol=<cident>
                          ( "AND" existcol=<cident> "TO" newcol=<cident> )*
                       ;
 
-<alterUserTypeStatement> ::= "ALTER" "TYPE" ut=<userTypeName>
+<alterUserTypeStatement> ::= "ALTER" "TYPE" ("IF" "EXISTS")? ut=<userTypeName>
                                <alterTypeInstructions>
                              ;
-<alterTypeInstructions> ::= "ADD" newcol=<cident> <storageType>
-                           | "RENAME" existcol=<cident> "TO" newcol=<cident>
+<alterTypeInstructions> ::= "ADD" ("IF" "NOT" "EXISTS")? newcol=<cident> <storageType>
+                           | "RENAME" ("IF" "EXISTS")? existcol=<cident> "TO" newcol=<cident>
                               ( "AND" existcol=<cident> "TO" newcol=<cident> )*
                            ;
 '''
@@ -1397,14 +1436,14 @@ syntax_rules += r'''
 def alter_table_col_completer(ctxt, cass):
     layout = get_table_meta(ctxt, cass)
     cols = [str(md) for md in layout.columns]
-    return map(maybe_escape_name, cols)
+    return list(map(maybe_escape_name, cols))
 
 
 @completer_for('alterTypeInstructions', 'existcol')
 def alter_type_field_completer(ctxt, cass):
     layout = get_ut_layout(ctxt, cass)
-    fields = [tuple[0] for tuple in layout]
-    return map(maybe_escape_name, fields)
+    fields = [atuple[0] for atuple in layout]
+    return list(map(maybe_escape_name, fields))
 
 
 explain_completion('alterInstructions', 'newcol', '<new_column_name>')
@@ -1412,7 +1451,7 @@ explain_completion('alterTypeInstructions', 'newcol', '<new_field_name>')
 
 
 syntax_rules += r'''
-<alterKeyspaceStatement> ::= "ALTER" wat=( "KEYSPACE" | "SCHEMA" ) ks=<alterableKeyspaceName>
+<alterKeyspaceStatement> ::= "ALTER" wat=( "KEYSPACE" | "SCHEMA" ) ("IF" "EXISTS")? ks=<alterableKeyspaceName>
                                  "WITH" <property> ( "AND" <property> )*
                            ;
 '''
@@ -1422,11 +1461,11 @@ syntax_rules += r'''
              ;
 
 <createUserStatement> ::= "CREATE" "USER" ( "IF" "NOT" "EXISTS" )? <username>
-                              ( "WITH" "PASSWORD" <stringLiteral> )?
+                              ( "WITH" ("HASHED")? "PASSWORD" <stringLiteral> )?
                               ( "SUPERUSER" | "NOSUPERUSER" )?
                         ;
 
-<alterUserStatement> ::= "ALTER" "USER" <username>
+<alterUserStatement> ::= "ALTER" "USER" ("IF" "EXISTS")? <username>
                               ( "WITH" "PASSWORD" <stringLiteral> )?
                               ( "SUPERUSER" | "NOSUPERUSER" )?
                        ;
@@ -1448,14 +1487,16 @@ syntax_rules += r'''
                               ( "WITH" <roleProperty> ("AND" <roleProperty>)*)?
                         ;
 
-<alterRoleStatement> ::= "ALTER" "ROLE" <rolename>
+<alterRoleStatement> ::= "ALTER" "ROLE" ("IF" "EXISTS")? <rolename>
                               ( "WITH" <roleProperty> ("AND" <roleProperty>)*)?
                        ;
 
-<roleProperty> ::= "PASSWORD" "=" <stringLiteral>
+<roleProperty> ::= (("HASHED")? "PASSWORD") "=" <stringLiteral>
                  | "OPTIONS" "=" <mapLiteral>
                  | "SUPERUSER" "=" <boolean>
                  | "LOGIN" "=" <boolean>
+                 | "ACCESS" "TO" "DATACENTERS" <setLiteral>
+                 | "ACCESS" "TO" "ALL" "DATACENTERS"
                  ;
 
 <dropRoleStatement> ::= "DROP" "ROLE" <rolename>
@@ -1493,7 +1534,7 @@ syntax_rules += r'''
                | "EXECUTE"
                ;
 
-<permissionExpr> ::= ( <permission> "PERMISSION"? )
+<permissionExpr> ::= ( [newpermission]=<permission> "PERMISSION"? ( "," [newpermission]=<permission> "PERMISSION"? )* )
                    | ( "ALL" "PERMISSIONS"? )
                    ;
 
@@ -1505,6 +1546,7 @@ syntax_rules += r'''
 
 <dataResource> ::= ( "ALL" "KEYSPACES" )
                  | ( "KEYSPACE" <keyspaceName> )
+                 | ( "ALL" "TABLES" "IN" "KEYSPACE" <keyspaceName> )
                  | ( "TABLE"? <columnFamilyName> )
                  ;
 
@@ -1527,6 +1569,16 @@ syntax_rules += r'''
 '''
 
 
+@completer_for('permissionExpr', 'newpermission')
+def permission_completer(ctxt, _):
+    new_permissions = set([permission.upper() for permission in ctxt.get_binding('newpermission')])
+    all_permissions = set([permission.arg for permission in ctxt.ruleset['permission'].arg])
+    suggestions = all_permissions - new_permissions
+    if len(suggestions) == 0:
+        return [Hint('No more permissions here.')]
+    return suggestions
+
+
 @completer_for('username', 'name')
 def username_name_completer(ctxt, cass):
     def maybe_quote(name):
@@ -1539,7 +1591,7 @@ def username_name_completer(ctxt, cass):
         return [Hint('<username>')]
 
     session = cass.session
-    return [maybe_quote(row.values()[0].replace("'", "''")) for row in session.execute("LIST USERS")]
+    return [maybe_quote(list(row.values())[0].replace("'", "''")) for row in session.execute("LIST USERS")]
 
 
 @completer_for('rolename', 'role')
@@ -1578,7 +1630,7 @@ def get_trigger_names(ctxt, cass):
 @completer_for('dropTriggerStatement', 'triggername')
 def drop_trigger_completer(ctxt, cass):
     names = get_trigger_names(ctxt, cass)
-    return map(maybe_escape_name, names)
+    return list(map(maybe_escape_name, names))
 
 
 # END SYNTAX/COMPLETION RULE DEFINITIONS
