@@ -16,29 +16,30 @@
 
 import re
 
-LIGHT = 010
+LIGHT = 0o10
 
-ansi_CSI = '\033['
+
+ansi_CSI = '\x1b['
 ansi_seq = re.compile(re.escape(ansi_CSI) + r'(?P<params>[\x20-\x3f]*)(?P<final>[\x40-\x7e])')
 ansi_cmd_SGR = 'm'  # set graphics rendition
 
 color_defs = (
     (000, 'k', 'black'),
-    (001, 'r', 'dark red'),
-    (002, 'g', 'dark green'),
-    (003, 'w', 'brown', 'dark yellow'),
-    (004, 'b', 'dark blue'),
-    (005, 'm', 'dark magenta', 'dark purple'),
-    (006, 'c', 'dark cyan'),
-    (007, 'n', 'light grey', 'light gray', 'neutral', 'dark white'),
-    (010, 'B', 'dark grey', 'dark gray', 'light black'),
-    (011, 'R', 'red', 'light red'),
-    (012, 'G', 'green', 'light green'),
-    (013, 'Y', 'yellow', 'light yellow'),
-    (014, 'B', 'blue', 'light blue'),
-    (015, 'M', 'magenta', 'purple', 'light magenta', 'light purple'),
-    (016, 'C', 'cyan', 'light cyan'),
-    (017, 'W', 'white', 'light white'),
+    (0o01, 'r', 'dark red'),
+    (0o02, 'g', 'dark green'),
+    (0o03, 'w', 'brown', 'dark yellow'),
+    (0o04, 'b', 'dark blue'),
+    (0o05, 'm', 'dark magenta', 'dark purple'),
+    (0o06, 'c', 'dark cyan'),
+    (0o07, 'n', 'light grey', 'light gray', 'neutral', 'dark white'),
+    (0o10, 'B', 'dark grey', 'dark gray', 'light black'),
+    (0o11, 'R', 'red', 'light red'),
+    (0o12, 'G', 'green', 'light green'),
+    (0o13, 'Y', 'yellow', 'light yellow'),
+    (0o14, 'B', 'blue', 'light blue'),
+    (0o15, 'M', 'magenta', 'purple', 'light magenta', 'light purple'),
+    (0o16, 'C', 'cyan', 'light cyan'),
+    (0o17, 'W', 'white', 'light white'),
 )
 
 colors_by_num = {}
@@ -61,7 +62,8 @@ for colordef in color_defs:
     for c in nameset:
         colors_by_name[c] = colorcode
 
-class ColoredChar:
+
+class ColoredChar(object):
     def __init__(self, c, colorcode):
         self.c = c
         self._colorcode = colorcode
@@ -76,8 +78,8 @@ class ColoredChar:
         return getattr(self.c, name)
 
     def ansi_color(self):
-        clr = str(30 + (07 & self._colorcode))
-        if self._colorcode & 010:
+        clr = str(30 + (0o7 & self._colorcode))
+        if self._colorcode & 0o10:
             clr = '1;' + clr
         return clr
 
@@ -100,11 +102,12 @@ class ColoredChar:
     def colortag(self):
         return lookup_letter_from_code(self._colorcode)
 
-class ColoredText:
+
+class ColoredText(object):
     def __init__(self, source=''):
-        if isinstance(source, basestring):
+        if isinstance(source, str):
             plain, colors = self.parse_ansi_colors(source)
-            self.chars = map(ColoredChar, plain, colors)
+            self.chars = list(map(ColoredChar, plain, colors))
         else:
             # expected that source is an iterable of ColoredChars (or duck-typed as such)
             self.chars = tuple(source)
@@ -148,8 +151,7 @@ class ColoredText:
 
     @staticmethod
     def parse_sgr_param(curclr, paramstr):
-        oldclr = curclr
-        args = map(int, paramstr.split(';'))
+        args = list(map(int, paramstr.split(';')))
         for a in args:
             if a == 0:
                 curclr = lookup_colorcode('neutral')
@@ -175,14 +177,18 @@ class ColoredText:
     def colortags(self):
         return ''.join([c.colortag() for c in self.chars])
 
+
 def lookup_colorcode(name):
     return colors_by_name[name]
+
 
 def lookup_colorname(code):
     return colors_by_num.get(code, 'Unknown-color-0%o' % code)
 
+
 def lookup_colorletter(letter):
     return colors_by_letter[letter]
+
 
 def lookup_letter_from_code(code):
     letr = letters_by_num.get(code, ' ')
