@@ -14,45 +14,51 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import sys
 import logging
-from itertools import izip
+import os
 from os.path import dirname, join, normpath
+import sys
+import unittest
 
 cqlshlog = logging.getLogger('test_cqlsh')
 
-try:
-    # a backport of python2.7 unittest features, so we can test against older
-    # pythons as necessary. python2.7 users who don't care about testing older
-    # versions need not install.
-    import unittest2 as unittest
-except ImportError:
-    import unittest
+test_dir = dirname(__file__)
+cassandra_dir = normpath(join(test_dir, '..', '..', '..'))
+cqlsh_dir = join(cassandra_dir, 'bin')
+path_to_cqlsh = join(cqlsh_dir, 'cqlsh.py')
 
-rundir = dirname(__file__)
-cqlshdir = normpath(join(rundir, '..', '..', '..', 'bin'))
-path_to_cqlsh = normpath(join(cqlshdir, 'cqlsh.py'))
-
-sys.path.append(cqlshdir)
-
-import cqlsh
-cql = cqlsh.cassandra.cluster.Cluster
-policy = cqlsh.cassandra.policies.RoundRobinPolicy()
-quote_name = cqlsh.cassandra.metadata.maybe_escape_name
+sys.path.append(cqlsh_dir)
 
 TEST_HOST = os.environ.get('CQL_TEST_HOST', '127.0.0.1')
 TEST_PORT = int(os.environ.get('CQL_TEST_PORT', 9042))
+TEST_USER = os.environ.get('CQL_TEST_USER', 'cassandra')
+TEST_PWD = os.environ.get('CQL_TEST_PWD')
+
 
 class BaseTestCase(unittest.TestCase):
     def assertNicelyFormattedTableHeader(self, line, msg=None):
-        return self.assertRegexpMatches(line, r'^ +\w+( +\| \w+)*\s*$', msg=msg)
+        return self.assertRegex(line, r'^ +\w+( +\| \w+)*\s*$', msg=msg)
 
     def assertNicelyFormattedTableRule(self, line, msg=None):
-        return self.assertRegexpMatches(line, r'^-+(\+-+)*\s*$', msg=msg)
+        return self.assertRegex(line, r'^-+(\+-+)*\s*$', msg=msg)
 
     def assertNicelyFormattedTableData(self, line, msg=None):
-        return self.assertRegexpMatches(line, r'^ .* \| ', msg=msg)
+        return self.assertRegex(line, r'^ .* \| ', msg=msg)
+
+    def assertRegex(self, text, regex, msg=None):
+        """Call assertRegexpMatches() if in Python 2"""
+        if hasattr(unittest.TestCase, 'assertRegex'):
+            return super().assertRegex(text, regex, msg)
+        else:
+            return self.assertRegexpMatches(text, regex, msg)
+
+    def assertNotRegex(self, text, regex, msg=None):
+        """Call assertNotRegexpMatches() if in Python 2"""
+        if hasattr(unittest.TestCase, 'assertNotRegex'):
+            return super().assertNotRegex(text, regex, msg)
+        else:
+            return self.assertNotRegexpMatches(text, regex, msg)
+
 
 def dedent(s):
     lines = [ln.rstrip() for ln in s.splitlines()]
@@ -62,5 +68,6 @@ def dedent(s):
     minspace = min(spaces if len(spaces) > 0 else (0,))
     return '\n'.join(line[minspace:] for line in lines)
 
+
 def at_a_time(i, num):
-    return izip(*([iter(i)] * num))
+    return zip(*([iter(i)] * num))
