@@ -14,44 +14,70 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import with_statement
 
 import contextlib
-import tempfile
+import io
 import os.path
-from .basecase import cql, cqlsh, cqlshlog, TEST_HOST, TEST_PORT, rundir, policy, quote_name
+import random
+import string
+
+from cassandra.cluster import Cluster
+from cassandra.metadata import maybe_escape_name as quote_name
+from cassandra.auth import PlainTextAuthProvider
+from cqlshlib.cql3handling import CqlRuleSet
+
+from .basecase import TEST_HOST, TEST_PORT, TEST_USER, TEST_PWD, cqlshlog, test_dir
 from .run_cqlsh import run_cqlsh, call_cqlsh
 
-test_keyspace_init = os.path.join(rundir, 'test_keyspace_init.cql')
+test_keyspace_init = os.path.join(test_dir, 'test_keyspace_init.cql')
+
 
 def get_cassandra_connection(cql_version=None):
-    conn = cql((TEST_HOST,), TEST_PORT, cql_version=cql_version, load_balancing_policy=policy)
+
+    auth_provider = PlainTextAuthProvider(username=TEST_USER, password=TEST_PWD)
+    conn = Cluster((TEST_HOST,), TEST_PORT, auth_provider=auth_provider, cql_version=cql_version)
+
     # until the cql lib does this for us
     conn.cql_version = cql_version
     return conn
 
+
 def get_cassandra_cursor(cql_version=None):
     return get_cassandra_connection(cql_version=cql_version).cursor()
 
+
 TEST_KEYSPACES_CREATED = []
+
 
 def get_keyspace():
     return None if len(TEST_KEYSPACES_CREATED) == 0 else TEST_KEYSPACES_CREATED[-1]
 
+
+_used_ks_names = set()
+
+
 def make_ks_name():
-    # abuse mktemp to get a quick random-ish name
-    return os.path.basename(tempfile.mktemp(prefix='CqlshTests_'))
+    def random_ks():
+        return 'cqlshtests_' + ''.join(random.choice(string.ascii_lowercase) for _ in range(10))
+
+    s = random_ks()
+    while s in _used_ks_names:
+        s = random_ks()
+    _used_ks_names.add(s)
+    return s
+
 
 def create_keyspace(cursor):
-    ksname = make_ks_name()
+    ksname = make_ks_name().lower()
     qksname = quote_name(ksname)
     cursor.execute('''
         CREATE KEYSPACE %s WITH replication =
             {'class': 'SimpleStrategy', 'replication_factor': 1};
-    ''' % quote_name(ksname))
+    ''' % qksname)
     cursor.execute('USE %s;' % qksname)
     TEST_KEYSPACES_CREATED.append(ksname)
     return ksname
+
 
 def split_cql_commands(source):
     ruleset = cql_rule_set()
@@ -61,14 +87,17 @@ def split_cql_commands(source):
 
     return [ruleset.cql_extract_orig(toks, source) for toks in statements if toks]
 
+
 def execute_cql_commands(cursor, source, logprefix='INIT: '):
     for cql in split_cql_commands(source):
-        cqlshlog.debug(logprefix + cql)
+        cqlshlog.debug((logprefix + cql).encode("utf-8"))
         cursor.execute(cql)
 
+
 def execute_cql_file(cursor, fname):
-    with open(fname) as f:
+    with io.open(fname, "r", encoding="utf-8") as f:
         return execute_cql_commands(cursor, f.read())
+
 
 def create_db():
     with cassandra_cursor(ks=None) as c:
@@ -76,9 +105,11 @@ def create_db():
         execute_cql_file(c, test_keyspace_init)
     return k
 
+
 def remove_db():
     with cassandra_cursor(ks=None) as c:
         c.execute('DROP KEYSPACE %s' % quote_name(TEST_KEYSPACES_CREATED.pop(-1)))
+
 
 @contextlib.contextmanager
 def cassandra_connection(cql_version=None):
@@ -94,7 +125,8 @@ def cassandra_connection(cql_version=None):
     try:
         yield conn
     finally:
-        conn.close()
+        conn.shutdown()
+
 
 @contextlib.contextmanager
 def cassandra_cursor(cql_version=None, ks=''):
@@ -116,17 +148,20 @@ def cassandra_cursor(cql_version=None, ks=''):
         c = conn.connect(ks)
         # increase default timeout to fix flacky tests, see CASSANDRA-12481
         c.default_timeout = 60.0
-        # if ks is not None:
-        #     c.execute('USE %s;' % quote_name(c, ks))
         yield c
     finally:
         conn.shutdown()
 
+
 def cql_rule_set():
-    return cqlsh.cql3handling.CqlRuleSet
+    return CqlRuleSet
 
-class DEFAULTVAL: pass
 
+class DEFAULTVAL:
+    pass
+
+
+__TEST__ = False
 def testrun_cqlsh(keyspace=DEFAULTVAL, **kwargs):
     # use a positive default sentinel so that keyspace=None can be used
     # to override the default behavior
@@ -134,7 +169,11 @@ def testrun_cqlsh(keyspace=DEFAULTVAL, **kwargs):
         keyspace = get_keyspace()
     return run_cqlsh(keyspace=keyspace, **kwargs)
 
+
+__TEST__ = False
 def testcall_cqlsh(keyspace=None, **kwargs):
     if keyspace is None:
         keyspace = get_keyspace()
+    if 'input' in kwargs.keys() and isinstance(kwargs['input'], str):
+        kwargs['input'] = kwargs['input'].encode('utf-8')
     return call_cqlsh(keyspace=keyspace, **kwargs)
