@@ -22,26 +22,44 @@ import re
 import contextlib
 import subprocess
 import signal
-import math
 from time import time
 from . import basecase
-from os.path import join, normpath
+from os.path import join
 
 
-def is_win():
-    return sys.platform in ("cygwin", "win32")
+import pty
+DEFAULT_PREFIX = os.linesep
 
-if is_win():
-    from winpty import WinPty
-    DEFAULT_PREFIX = ''
-else:
-    import pty
-    DEFAULT_PREFIX = os.linesep
-
-DEFAULT_CQLSH_PROMPT = DEFAULT_PREFIX + '(\S+@)?cqlsh(:\S+)?> '
+DEFAULT_CQLSH_PROMPT = DEFAULT_PREFIX + r'(\S+@)?cqlsh(:\S+)?> '
 DEFAULT_CQLSH_TERM = 'xterm'
 
+try:
+    Pattern = re._pattern_type
+except AttributeError:
+    # Python 3.7+
+    Pattern = re.Pattern
+
+
+def get_smm_sequence(term='xterm'):
+    """
+    Return the set meta mode (smm) sequence, if any.
+    On more recent Linux systems, xterm emits the smm sequence
+    before each prompt.
+    """
+    result = ''
+    tput_proc = subprocess.Popen(['tput', '-T{}'.format(term), 'smm'], stdout=subprocess.PIPE)
+    tput_stdout = tput_proc.communicate()[0]
+    if (tput_stdout and (tput_stdout != b'')):
+        result = tput_stdout
+        if isinstance(result, bytes):
+            result = result.decode("utf-8")
+    return result
+
+
+DEFAULT_SMM_SEQUENCE = get_smm_sequence()
+
 cqlshlog = basecase.cqlshlog
+
 
 def set_controlling_pty(master, slave):
     os.setsid()
@@ -51,6 +69,7 @@ def set_controlling_pty(master, slave):
     if slave > 2:
         os.close(slave)
     os.close(os.open(os.ttyname(1), os.O_RDWR))
+
 
 @contextlib.contextmanager
 def raising_signal(signum, exc):
@@ -67,11 +86,13 @@ def raising_signal(signum, exc):
     finally:
         signal.signal(signum, oldhandlr)
 
+
 class TimeoutError(Exception):
     pass
 
+
 @contextlib.contextmanager
-def timing_out_itimer(seconds):
+def timing_out(seconds):
     if seconds is None:
         yield
         return
@@ -85,46 +106,16 @@ def timing_out_itimer(seconds):
         finally:
             signal.setitimer(signal.ITIMER_REAL, 0)
 
-@contextlib.contextmanager
-def timing_out_alarm(seconds):
-    if seconds is None:
-        yield
-        return
-    with raising_signal(signal.SIGALRM, TimeoutError):
-        oldval = signal.alarm(int(math.ceil(seconds)))
-        if oldval != 0:
-            signal.alarm(oldval)
-            raise RuntimeError("SIGALRM already in use")
-        try:
-            yield
-        finally:
-            signal.alarm(0)
-
-if is_win():
-    try:
-        import eventlet
-    except ImportError, e:
-        sys.exit("evenlet library required to run cqlshlib tests on Windows")
-
-    def timing_out(seconds):
-        return eventlet.Timeout(seconds, TimeoutError)
-else:
-    # setitimer is new in 2.6, but it's still worth supporting, for potentially
-    # faster tests because of sub-second resolution on timeouts.
-    if hasattr(signal, 'setitimer'):
-        timing_out = timing_out_itimer
-    else:
-        timing_out = timing_out_alarm
 
 def noop(*a):
     pass
+
 
 class ProcRunner:
     def __init__(self, path, tty=True, env=None, args=()):
         self.exe_path = path
         self.args = args
-        self.tty = bool(tty)
-        self.realtty = self.tty and not is_win()
+        self.tty = tty
         if env is None:
             env = {}
         self.env = env
@@ -137,7 +128,7 @@ class ProcRunner:
         stdin = stdout = stderr = None
         cqlshlog.info("Spawning %r subprocess with args: %r and env: %r"
                       % (self.exe_path, self.args, self.env))
-        if self.realtty:
+        if self.tty:
             masterfd, slavefd = pty.openpty()
             preexec = (lambda: set_controlling_pty(masterfd, slavefd))
             self.proc = subprocess.Popen((self.exe_path,) + tuple(self.args),
@@ -155,15 +146,11 @@ class ProcRunner:
                                          env=self.env, stdin=stdin, stdout=stdout,
                                          stderr=stderr, bufsize=0, close_fds=False)
             self.send = self.send_pipe
-            if self.tty:
-                self.winpty = WinPty(self.proc.stdout)
-                self.read = self.read_winpty
-            else:
-                self.read = self.read_pipe
+            self.read = self.read_pipe
 
     def close(self):
         cqlshlog.info("Closing %r subprocess." % (self.exe_path,))
-        if self.realtty:
+        if self.tty:
             os.close(self.childpty)
         else:
             self.proc.stdin.close()
@@ -171,23 +158,28 @@ class ProcRunner:
         return self.proc.wait()
 
     def send_tty(self, data):
+        if not isinstance(data, bytes):
+            data = data.encode("utf-8")
         os.write(self.childpty, data)
 
     def send_pipe(self, data):
         self.proc.stdin.write(data)
 
     def read_tty(self, blksize, timeout=None):
-        return os.read(self.childpty, blksize)
+        buf = os.read(self.childpty, blksize)
+        if isinstance(buf, bytes):
+            buf = buf.decode("utf-8")
+        return buf
 
     def read_pipe(self, blksize, timeout=None):
-        return self.proc.stdout.read(blksize)
-
-    def read_winpty(self, blksize, timeout=None):
-        return self.winpty.read(blksize, timeout)
+        buf = self.proc.stdout.read(blksize)
+        if isinstance(buf, bytes):
+            buf = buf.decode("utf-8")
+        return buf
 
     def read_until(self, until, blksize=4096, timeout=None,
-                   flags=0, ptty_timeout=None):
-        if not isinstance(until, re._pattern_type):
+                   flags=0, ptty_timeout=None, replace=[]):
+        if not isinstance(until, Pattern):
             until = re.compile(until, flags)
 
         cqlshlog.debug("Searching for %r" % (until.pattern,))
@@ -196,6 +188,9 @@ class ProcRunner:
         with timing_out(timeout):
             while True:
                 val = self.read(blksize, ptty_timeout)
+                for replace_target in replace:
+                    if (replace_target != ''):
+                        val = val.replace(replace_target, '')
                 cqlshlog.debug("read %r from subproc" % (val,))
                 if val == '':
                     raise EOFError("'until' pattern %r not found" % (until.pattern,))
@@ -231,38 +226,33 @@ class ProcRunner:
             curtime = time()
         return got
 
+
 class CqlshRunner(ProcRunner):
     def __init__(self, path=None, host=None, port=None, keyspace=None, cqlver=None,
-                 args=(), prompt=DEFAULT_CQLSH_PROMPT, env=None,
-                 win_force_colors=True, tty=True, **kwargs):
+                 args=(), prompt=DEFAULT_CQLSH_PROMPT, env=None, tty=True, **kwargs):
         if path is None:
-            cqlsh_bin = 'cqlsh'
-            if is_win():
-                cqlsh_bin = 'cqlsh.bat'
-            path = normpath(join(basecase.cqlshdir, cqlsh_bin))
+            path = join(basecase.cqlsh_dir, 'cqlsh')
         if host is None:
             host = basecase.TEST_HOST
         if port is None:
             port = basecase.TEST_PORT
         if env is None:
             env = {}
-        if is_win():
-            env['PYTHONUNBUFFERED'] = '1'
-            env.update(os.environ.copy())
         env.setdefault('TERM', 'xterm')
         env.setdefault('CQLSH_NO_BUNDLED', os.environ.get('CQLSH_NO_BUNDLED', ''))
         env.setdefault('PYTHONPATH', os.environ.get('PYTHONPATH', ''))
+        coverage = False
+        if ('CQLSH_COVERAGE' in env.keys()):
+            coverage = True
         args = tuple(args) + (host, str(port))
         if cqlver is not None:
             args += ('--cqlversion', str(cqlver))
         if keyspace is not None:
-            args += ('--keyspace', keyspace)
-        if tty and is_win():
-            args += ('--tty',)
-            args += ('--encoding', 'utf-8')
-            if win_force_colors:
-                args += ('--color',)
+            args += ('--keyspace', keyspace.lower())
+        if coverage:
+            args += ('--coverage',)
         self.keyspace = keyspace
+        env.setdefault('CQLSH_PYTHON', sys.executable)  # run with the same interpreter as the test
         ProcRunner.__init__(self, path, tty=tty, args=args, env=env, **kwargs)
         self.prompt = prompt
         if self.prompt is None:
@@ -270,8 +260,8 @@ class CqlshRunner(ProcRunner):
         else:
             self.output_header = self.read_to_next_prompt()
 
-    def read_to_next_prompt(self):
-        return self.read_until(self.prompt, timeout=10.0, ptty_timeout=3)
+    def read_to_next_prompt(self, timeout=10.0):
+        return self.read_until(self.prompt, timeout=timeout, ptty_timeout=3, replace=[DEFAULT_SMM_SEQUENCE])
 
     def read_up_to_timeout(self, timeout, blksize=4096):
         output = ProcRunner.read_up_to_timeout(self, timeout, blksize=blksize)
@@ -287,7 +277,7 @@ class CqlshRunner(ProcRunner):
         output = output.replace(' \r', '')
         output = output.replace('\r', '')
         output = output.replace(' \b', '')
-        if self.realtty:
+        if self.tty:
             echo, output = output.split('\n', 1)
             assert echo == cmd, "unexpected echo %r instead of %r" % (echo, cmd)
         try:
@@ -296,11 +286,13 @@ class CqlshRunner(ProcRunner):
             promptline = output
             output = ''
         assert re.match(self.prompt, DEFAULT_PREFIX + promptline), \
-                'last line of output %r does not match %r?' % (promptline, self.prompt)
+            'last line of output %r does not match %r?' % (promptline, self.prompt)
         return output + '\n'
+
 
 def run_cqlsh(**kwargs):
     return contextlib.closing(CqlshRunner(**kwargs))
+
 
 def call_cqlsh(**kwargs):
     kwargs.setdefault('prompt', None)
@@ -309,4 +301,6 @@ def call_cqlsh(**kwargs):
     c = CqlshRunner(**kwargs)
     output, _ = c.proc.communicate(proginput)
     result = c.close()
+    if isinstance(output, bytes):
+        output = output.decode("utf-8")
     return output, result
