@@ -15,37 +15,52 @@
 # limitations under the License.
 
 # to configure behavior, define $CQL_TEST_HOST to the destination address
-# for Thrift connections, and $CQL_TEST_PORT to the associated port.
+# and $CQL_TEST_PORT to the associated port.
 
-from __future__ import with_statement
-
+import locale
+import os
 import re
-from itertools import izip
-from .basecase import (BaseTestCase, cqlshlog, dedent, at_a_time, cqlsh,
-                       TEST_HOST, TEST_PORT)
-from .cassconnect import (get_keyspace, testrun_cqlsh, testcall_cqlsh,
-                          cassandra_cursor, split_cql_commands, quote_name)
-from .ansi_colors import (ColoredText, lookup_colorcode, lookup_colorname,
-                          lookup_colorletter, ansi_seq)
-import unittest
-import sys
+
+from .basecase import (BaseTestCase, TEST_HOST, TEST_PORT,
+                       at_a_time, cqlshlog, dedent)
+from .cassconnect import (cassandra_cursor, create_db, get_keyspace,
+                          quote_name, remove_db, split_cql_commands,
+                          testcall_cqlsh, testrun_cqlsh)
+from .ansi_colors import (ColoredText, ansi_seq, lookup_colorcode,
+                          lookup_colorname, lookup_colorletter)
 
 CONTROL_C = '\x03'
 CONTROL_D = '\x04'
 
+
 class TestCqlshOutput(BaseTestCase):
 
+    @classmethod
+    def setUpClass(cls):
+        create_db()
+
+    @classmethod
+    def tearDownClass(cls):
+        remove_db()
+
     def setUp(self):
-        pass
+        env = os.environ.copy()
+        env['COLUMNS'] = '100000'
+        # carry forward or override locale LC_CTYPE for UTF-8 encoding
+        if (locale.getpreferredencoding() != 'UTF-8'):
+            env['LC_CTYPE'] = 'en_US.utf8'
+        else:
+            env['LC_CTYPE'] = os.environ.get('LC_CTYPE', 'en_US.utf8')
+        self.default_env = env
 
     def tearDown(self):
         pass
 
     def assertNoHasColors(self, text, msg=None):
-        self.assertNotRegexpMatches(text, ansi_seq, msg='ANSI CSI sequence found in %r' % text)
+        self.assertNotRegex(text, ansi_seq, msg='ANSI CSI sequence found in %r' % text)
 
     def assertHasColors(self, text, msg=None):
-        self.assertRegexpMatches(text, ansi_seq, msg=msg)
+        self.assertRegex(text, ansi_seq, msg=msg)
 
     def assertColored(self, coloredtext, colorname):
         wanted_colorcode = lookup_colorcode(colorname)
@@ -57,7 +72,7 @@ class TestCqlshOutput(BaseTestCase):
                                      % (coloredtext, num, lookup_colorname(ccolor), colorname))
 
     def assertColorFromTags(self, coloredtext, tags):
-        for (char, tag) in izip(coloredtext, tags):
+        for (char, tag) in zip(coloredtext, tags):
             if char.isspace():
                 continue
             if tag.isspace():
@@ -67,12 +82,14 @@ class TestCqlshOutput(BaseTestCase):
                                  'Actually got:      %s\ncolor code:        %s'
                                  % (tags, coloredtext.colored_version(), coloredtext.colortags()))
 
-    def assertQueriesGiveColoredOutput(self, queries_and_expected_outputs, **kwargs):
+    def assertQueriesGiveColoredOutput(self, queries_and_expected_outputs, env=None, **kwargs):
         """
         Allow queries and expected output to be specified in structured tuples,
         along with expected color information.
         """
-        with testrun_cqlsh(tty=True, **kwargs) as c:
+        if env is None:
+            env = self.default_env
+        with testrun_cqlsh(tty=True, env=env, **kwargs) as c:
             for query, expected in queries_and_expected_outputs:
                 cqlshlog.debug('Testing %r' % (query,))
                 output = c.cmd_and_response(query).lstrip("\r\n")
@@ -83,11 +100,25 @@ class TestCqlshOutput(BaseTestCase):
                     self.assertEqual(outputline.plain().rstrip(), plain)
                     self.assertColorFromTags(outputline, colorcodes)
 
+    def strip_read_repair_chance(self, describe_statement):
+        """
+        Remove read_repair_chance and dclocal_read_repair_chance options
+        from output of DESCRIBE statements. The resulting string may be
+        reused as a CREATE statement.
+        Useful after CASSANDRA-13910, which removed read_repair_chance
+        options from CREATE statements but did not remove them completely
+        from the system.
+        """
+        describe_statement = re.sub(r"( AND)? (dclocal_)?read_repair_chance = [\d\.]+", "", describe_statement)
+        describe_statement = re.sub(r"WITH[\s]*;", "", describe_statement)
+        return describe_statement
+
     def test_no_color_output(self):
+        env = self.default_env
         for termname in ('', 'dumb', 'vt100'):
             cqlshlog.debug('TERM=%r' % termname)
-            with testrun_cqlsh(tty=True, env={'TERM': termname},
-                               win_force_colors=False) as c:
+            env['TERM'] = termname
+            with testrun_cqlsh(tty=True, env=env) as c:
                 c.send('select * from has_all_types;\n')
                 self.assertNoHasColors(c.read_to_next_prompt())
                 c.send('select count(*) from has_all_types;\n')
@@ -96,15 +127,17 @@ class TestCqlshOutput(BaseTestCase):
                 self.assertNoHasColors(c.read_to_next_prompt())
 
     def test_no_prompt_or_colors_output(self):
+        env = self.default_env
         for termname in ('', 'dumb', 'vt100', 'xterm'):
             cqlshlog.debug('TERM=%r' % termname)
+            env['TERM'] = termname
             query = 'select * from has_all_types limit 1;'
-            output, result = testcall_cqlsh(prompt=None, env={'TERM': termname},
+            output, result = testcall_cqlsh(prompt=None, env=env,
                                             tty=False, input=query + '\n')
             output = output.splitlines()
             for line in output:
                 self.assertNoHasColors(line)
-                self.assertNotRegexpMatches(line, r'^cqlsh\S*>')
+                self.assertNotRegex(line, r'^cqlsh\S*>')
             self.assertEqual(len(output), 6,
                              msg='output: %r' % '\n'.join(output))
             self.assertEqual(output[0], '')
@@ -115,9 +148,12 @@ class TestCqlshOutput(BaseTestCase):
             self.assertEqual(output[5].strip(), '(1 rows)')
 
     def test_color_output(self):
+        env = self.default_env
         for termname in ('xterm', 'unknown-garbage'):
             cqlshlog.debug('TERM=%r' % termname)
-            with testrun_cqlsh(tty=True, env={'TERM': termname}) as c:
+            env['TERMNAME'] = termname
+            env['TERM'] = termname
+            with testrun_cqlsh(tty=True, env=env) as c:
                 c.send('select * from has_all_types;\n')
                 self.assertHasColors(c.read_to_next_prompt())
                 c.send('select count(*) from has_all_types;\n')
@@ -349,6 +385,8 @@ class TestCqlshOutput(BaseTestCase):
         ))
 
     def test_timestamp_output(self):
+        env = self.default_env
+        env['TZ'] = 'Etc/UTC'
         self.assertQueriesGiveColoredOutput((
             ('''select timestampcol from has_all_types where num = 0;''', """
              timestampcol
@@ -362,9 +400,10 @@ class TestCqlshOutput(BaseTestCase):
             (1 rows)
             nnnnnnnn
             """),
-        ), env={'TZ': 'Etc/UTC'})
+        ), env=env)
         try:
             import pytz  # test only if pytz is available on PYTHONPATH
+            env['TZ'] = 'America/Sao_Paulo'
             self.assertQueriesGiveColoredOutput((
                 ('''select timestampcol from has_all_types where num = 0;''', """
                  timestampcol
@@ -378,7 +417,7 @@ class TestCqlshOutput(BaseTestCase):
                 (1 rows)
                 nnnnnnnn
                 """),
-            ), env={'TZ': 'America/Sao_Paulo'})
+            ), env=env)
         except ImportError:
             pass
 
@@ -470,8 +509,10 @@ class TestCqlshOutput(BaseTestCase):
         # terminals, but the color-checking machinery here will still treat
         # it as one character, so those won't seem to line up visually either.
 
+        env = self.default_env
+        env['LANG'] = 'en_US.UTF-8'
         self.assertQueriesGiveColoredOutput((
-            ("select * from utf8_with_special_chars where k in (0, 1, 2, 3, 4, 5, 6);", u"""
+            ("select * from utf8_with_special_chars where k in (0, 1, 2, 3, 4, 5, 6);", """
              k | val
              R   MMM
             ---+-------------------------------
@@ -494,8 +535,8 @@ class TestCqlshOutput(BaseTestCase):
 
             (7 rows)
             nnnnnnnn
-            """.encode('utf-8')),
-        ), env={'LANG': 'en_US.UTF-8'})
+            """),
+        ), env=env)
 
     def test_blob_output(self):
         self.assertQueriesGiveColoredOutput((
@@ -516,11 +557,11 @@ class TestCqlshOutput(BaseTestCase):
 
             (4 rows)
             nnnnnnnn
-            """),
+            """, ),
         ))
 
     def test_prompt(self):
-        with testrun_cqlsh(tty=True, keyspace=None) as c:
+        with testrun_cqlsh(tty=True, keyspace=None, env=self.default_env) as c:
             self.assertTrue(c.output_header.splitlines()[-1].endswith('cqlsh> '))
 
             c.send('\n')
@@ -540,7 +581,7 @@ class TestCqlshOutput(BaseTestCase):
             outputlines = c.read_to_next_prompt().splitlines()
 
             start_index = 0
-            if c.realtty:
+            if c.tty:
                 self.assertEqual(outputlines[start_index], 'use NONEXISTENTKEYSPACE;')
                 start_index = 1
 
@@ -552,7 +593,7 @@ class TestCqlshOutput(BaseTestCase):
                              "RRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRRR")
 
     def test_describe_keyspace_output(self):
-        with testrun_cqlsh(tty=True) as c:
+        with testrun_cqlsh(tty=True, env=self.default_env) as c:
             ks = get_keyspace()
             qks = quote_name(ks)
             for cmd in ('describe keyspace', 'desc keyspace'):
@@ -572,19 +613,20 @@ class TestCqlshOutput(BaseTestCase):
             with cassandra_cursor() as curs:
                 try:
                     for stmt in statements:
+                        stmt = self.strip_read_repair_chance(stmt)
                         cqlshlog.debug('TEST EXEC: %s' % stmt)
                         curs.execute(stmt)
                 finally:
                     curs.execute('use system')
                     if do_drop:
-                        curs.execute('drop keyspace %s' % quote_name(new_ks_name))
+                        curs.execute('drop keyspace {}'.format(new_ks_name))
 
     def check_describe_keyspace_output(self, output, qksname):
         expected_bits = [r'(?im)^CREATE KEYSPACE %s WITH\b' % re.escape(qksname),
                          r';\s*$',
                          r'\breplication = {\'class\':']
         for expr in expected_bits:
-            self.assertRegexpMatches(output, expr)
+            self.assertRegex(output, expr)
 
     def test_describe_columnfamily_output(self):
         # we can change these to regular expressions if/when it makes sense
@@ -592,7 +634,6 @@ class TestCqlshOutput(BaseTestCase):
 
         # note columns are now comparator-ordered instead of original-order.
         table_desc3 = dedent("""
-
             CREATE TABLE %s.has_all_types (
                 num int PRIMARY KEY,
                 asciicol ascii,
@@ -610,43 +651,41 @@ class TestCqlshOutput(BaseTestCase):
                 uuidcol uuid,
                 varcharcol text,
                 varintcol varint
-            ) WITH bloom_filter_fp_chance = 0.01
+            ) WITH additional_write_policy = '99p'
+                AND bloom_filter_fp_chance = 0.01
                 AND caching = {'keys': 'ALL', 'rows_per_partition': 'NONE'}
                 AND cdc = false
                 AND comment = ''
                 AND compaction = {'class': 'org.apache.cassandra.db.compaction.SizeTieredCompactionStrategy', 'max_threshold': '32', 'min_threshold': '4'}
-                AND compression = {'chunk_length_in_kb': '64', 'class': 'org.apache.cassandra.io.compress.LZ4Compressor'}
+                AND compression = {'chunk_length_in_kb': '16', 'class': 'org.apache.cassandra.io.compress.LZ4Compressor'}
+                AND memtable = 'default'
                 AND crc_check_chance = 1.0
-                AND dclocal_read_repair_chance = 0.1
                 AND default_time_to_live = 0
+                AND extensions = {}
                 AND gc_grace_seconds = 864000
                 AND max_index_interval = 2048
                 AND memtable_flush_period_in_ms = 0
                 AND min_index_interval = 128
-                AND read_repair_chance = 0.0
-                AND speculative_retry = '99PERCENTILE';
+                AND read_repair = 'BLOCKING'
+                AND speculative_retry = '99p';""" % quote_name(get_keyspace()))
 
-        """ % quote_name(get_keyspace()))
-
-        with testrun_cqlsh(tty=True) as c:
+        with testrun_cqlsh(tty=True, env=self.default_env) as c:
             for cmdword in ('describe table', 'desc columnfamily'):
                 for semicolon in (';', ''):
                     output = c.cmd_and_response('%s has_all_types%s' % (cmdword, semicolon))
                     self.assertNoHasColors(output)
-                    self.assertSequenceEqual(output.split('\n'), table_desc3.split('\n'))
+                    self.assertSequenceEqual(dedent(output).split('\n'), table_desc3.split('\n'))
 
     def test_describe_columnfamilies_output(self):
         output_re = r'''
-            \n
-            Keyspace [ ] (?P<ksname> \S+ ) \n
+            \n? Keyspace [ ] (?P<ksname> \S+ ) \n
             -----------* \n
-            (?P<cfnames> .*? )
-            \n
-        '''
+            (?P<cfnames> ( ( ["']?\w+["']? [^\w\n]* )+ \n )* )
+            '''
 
         ks = get_keyspace()
 
-        with testrun_cqlsh(tty=True, keyspace=None) as c:
+        with testrun_cqlsh(tty=True, keyspace=None, env=self.default_env) as c:
 
             # when not in a keyspace
             for cmdword in ('DESCRIBE COLUMNFAMILIES', 'desc tables'):
@@ -654,7 +693,7 @@ class TestCqlshOutput(BaseTestCase):
                     ksnames = []
                     output = c.cmd_and_response(cmdword + semicolon)
                     self.assertNoHasColors(output)
-                    self.assertRegexpMatches(output, '(?xs) ^ ( %s )+ $' % output_re)
+                    self.assertRegex(output, '(?xs) ^ ( %s )+ $' % output_re)
 
                     for section in re.finditer('(?xs)' + output_re, output):
                         ksname = section.group('ksname')
@@ -686,24 +725,25 @@ class TestCqlshOutput(BaseTestCase):
             \n
             Cluster: [ ] (?P<clustername> .* ) \n
             Partitioner: [ ] (?P<partitionername> .* ) \n
+            Snitch: [ ] (?P<snitchname> .* ) \n
             \n
         '''
 
         ringinfo_re = r'''
             Range[ ]ownership: \n
             (
-              [ ] .*? [ ][ ] \[ ( \d+ \. ){3} \d+ \] \n
+              [ ] .*? [ ][ ] \[ .*? / ( \d+ \. ){3} \d+ : \d+ \] \n
             )+
             \n
         '''
 
-        with testrun_cqlsh(tty=True, keyspace=None) as c:
+        with testrun_cqlsh(tty=True, keyspace=None, env=self.default_env) as c:
 
             # not in a keyspace
             for semicolon in ('', ';'):
                 output = c.cmd_and_response('describe cluster' + semicolon)
                 self.assertNoHasColors(output)
-                self.assertRegexpMatches(output, output_re + '$')
+                self.assertRegex(output, output_re + '$')
 
             c.send('USE %s;\n' % quote_name(get_keyspace()))
             c.read_to_next_prompt()
@@ -711,32 +751,32 @@ class TestCqlshOutput(BaseTestCase):
             for semicolon in ('', ';'):
                 output = c.cmd_and_response('describe cluster' + semicolon)
                 self.assertNoHasColors(output)
-                self.assertRegexpMatches(output, output_re + ringinfo_re + '$')
+                self.assertRegex(output, output_re + ringinfo_re + '$')
 
     def test_describe_schema_output(self):
-        with testrun_cqlsh(tty=True) as c:
+        with testrun_cqlsh(tty=True, env=self.default_env) as c:
             for semicolon in ('', ';'):
                 output = c.cmd_and_response('desc full schema' + semicolon)
                 self.assertNoHasColors(output)
-                self.assertRegexpMatches(output, '^\nCREATE KEYSPACE')
-                self.assertIn("\nCREATE KEYSPACE system WITH replication = {'class': 'LocalStrategy'}  AND durable_writes = true;\n",
+                # Since CASSANDRA-7622 'DESC FULL SCHEMA' also shows all VIRTUAL keyspaces
+                self.assertIn('VIRTUAL KEYSPACE system_virtual_schema', output)
+                self.assertIn("\nCREATE KEYSPACE system_auth WITH replication = {'class': 'SimpleStrategy', 'replication_factor': '1'}  AND durable_writes = true;\n",
                               output)
-                self.assertRegexpMatches(output, ';\s*$')
+                self.assertRegex(output, r'.*\s*$')
 
     def test_show_output(self):
-        with testrun_cqlsh(tty=True) as c:
+        with testrun_cqlsh(tty=True, env=self.default_env) as c:
             output = c.cmd_and_response('show version;')
-            self.assertRegexpMatches(output,
-                    '^\[cqlsh \S+ \| Cassandra \S+ \| CQL spec \S+ \| Native protocol \S+\]$')
+            self.assertRegex(output,
+                    r'^\[cqlsh \S+ \| Cassandra \S+ \| CQL spec \S+ \| Native protocol \S+\]$')
 
             output = c.cmd_and_response('show host;')
             self.assertHasColors(output)
-            self.assertRegexpMatches(output, '^Connected to .* at %s:%d\.$'
+            self.assertRegex(output, '^Connected to .* at %s:%d$'
                                              % (re.escape(TEST_HOST), TEST_PORT))
 
-    @unittest.skipIf(sys.platform == "win32", 'EOF signaling not supported on Windows')
     def test_eof_prints_newline(self):
-        with testrun_cqlsh(tty=True) as c:
+        with testrun_cqlsh(tty=True, env=self.default_env) as c:
             c.send(CONTROL_D)
             out = c.read_lines(1)[0].replace('\r', '')
             self.assertEqual(out, '\n')
@@ -746,18 +786,17 @@ class TestCqlshOutput(BaseTestCase):
 
     def test_exit_prints_no_newline(self):
         for semicolon in ('', ';'):
-            with testrun_cqlsh(tty=True) as c:
+            with testrun_cqlsh(tty=True, env=self.default_env) as c:
                 cmd = 'exit%s\n' % semicolon
                 c.send(cmd)
-                if c.realtty:
-                    out = c.read_lines(1)[0].replace('\r', '')
-                    self.assertEqual(out, cmd)
+                out = c.read_lines(1)[0].replace('\r', '')
+                self.assertEqual(out, cmd)
                 with self.assertRaises(BaseException) as cm:
                     c.read_lines(1)
                 self.assertIn(type(cm.exception), (EOFError, OSError))
 
     def test_help_types(self):
-        with testrun_cqlsh(tty=True) as c:
+        with testrun_cqlsh(tty=True, env=self.default_env) as c:
             c.cmd_and_response('help types')
 
     def test_help(self):
@@ -801,6 +840,7 @@ class TestCqlshOutput(BaseTestCase):
             nnnnnnnn
             """),
         ))
+
         self.assertQueriesGiveColoredOutput((
             ("select phone_numbers from users;", r"""
              phone_numbers
@@ -833,6 +873,7 @@ class TestCqlshOutput(BaseTestCase):
             nnnnnnnn
             """),
         ))
+
         self.assertQueriesGiveColoredOutput((
             ("select tags from songs;", r"""
              tags
@@ -862,3 +903,13 @@ class TestCqlshOutput(BaseTestCase):
             nnnnnnnn
             """),
         ))
+
+    def test_expanded_output_counts_past_page(self):
+        query = "PAGING 5; EXPAND ON; SELECT * FROM twenty_rows_table;"
+        output, result = testcall_cqlsh(prompt=None, env=self.default_env,
+                                        tty=False, input=query)
+        self.assertEqual(0, result)
+        # format is "@ Row 1"
+        row_headers = [s for s in output.splitlines() if "@ Row" in s]
+        row_ids = [int(s.split(' ')[2]) for s in row_headers]
+        self.assertEqual([i for i in range(1, 21)], row_ids)
